refactor(console):; move properties component in dedicated directory
