+++ gravitee-apim-console-webui/src/management/api/proxy/properties/properties/api-properties.component.html
