+++ gravitee-apim-console-webui/src/management/api/proxy/properties/properties/api-properties.component.scss
@@ -1,7 +1,7 @@
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
-@use '../../../../scss/gio-layout' as gio-layout;
+@use '../../../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-margin-container;
