+++ gravitee-apim-console-webui/src/management/api/proxy/properties/properties/properties-add-dialog/properties-add-dialog.component.html
