+++ gravitee-apim-console-webui/src/management/api/proxy/properties/properties/properties-import-dialog/properties-import-dialog.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 .title {
   display: flex;
